Moved permission request from plugin initialization to view initialization.
