+++ android/src/main/kotlin/net/touchcapture/qr/flutterqr/FlutterQrPlugin.kt
@@ -1,25 +1,15 @@
 package net.touchcapture.qr.flutterqr
 
-import android.Manifest
 import android.app.Activity
-import android.content.pm.PackageManager
-import android.os.Build
 import androidx.annotation.NonNull
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.BinaryMessenger
-import io.flutter.plugin.common.MethodCall
-import io.flutter.plugin.common.MethodChannel.MethodCallHandler
-import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry
 import io.flutter.plugin.platform.PlatformViewRegistry
 
-
-class FlutterQrPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
-
-    var cameraPermissionContinuation: Runnable? = null
-    var requestingPermission = false
+class FlutterQrPlugin : FlutterPlugin, ActivityAware {
 
     /** Plugin registration embedding v1 */
     companion object {
@@ -31,7 +21,6 @@ class FlutterQrPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
     private fun onAttachedToV1(registrar: PluginRegistry.Registrar) {
         registrar.addRequestPermissionsResultListener(CameraRequestPermissionsListener())
-        checkAndRequestPermission(null)
         onAttachedToEngines(registrar.platformViewRegistry(), registrar.messenger(), registrar.activity())
     }
 
@@ -53,16 +42,9 @@ class FlutterQrPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
                         "net.touchcapture.qr.flutterqr/qrview", QRViewFactory(messenger))
     }
 
-    override fun onMethodCall(call: MethodCall, result: Result) {
-        when (call.method) {
-            "checkAndRequestPermission" -> checkAndRequestPermission(result)
-        }
-    }
-
     override fun onAttachedToActivity(activityPluginBinding: ActivityPluginBinding) {
         Shared.activity = activityPluginBinding.activity
         activityPluginBinding.addRequestPermissionsResultListener(CameraRequestPermissionsListener())
-        checkAndRequestPermission(null)
     }
 
     override fun onDetachedFromActivityForConfigChanges() {
@@ -75,47 +57,5 @@ class FlutterQrPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
     override fun onDetachedFromActivity() {
         Shared.activity = null
-    }
-
-    private inner class CameraRequestPermissionsListener : PluginRegistry.RequestPermissionsResultListener {
-        override fun onRequestPermissionsResult(id: Int, permissions: Array<String>, grantResults: IntArray): Boolean {
-            if (id == QRView.CAMERA_REQUEST_ID && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                cameraPermissionContinuation?.run()
-                return true
-            }
-            return false
-        }
-    }
-
-    private fun hasCameraPermission(): Boolean {
-        return Build.VERSION.SDK_INT < Build.VERSION_CODES.M ||
-                Shared.activity?.checkSelfPermission(Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED
-    }
-
-    private fun checkAndRequestPermission(result: Result?) {
-        if (cameraPermissionContinuation != null) {
-            result?.error("cameraPermission", "Camera permission request ongoing", null)
-        }
-
-        cameraPermissionContinuation = Runnable {
-            cameraPermissionContinuation = null
-            if (!hasCameraPermission()) {
-                result?.error(
-                        "cameraPermission", "MediaRecorderCamera permission not granted", null)
-                return@Runnable
-            }
-        }
-
-        requestingPermission = false
-        if (hasCameraPermission()) {
-            cameraPermissionContinuation?.run()
-        } else {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                requestingPermission = true
-                Shared.activity?.requestPermissions(
-                        arrayOf(Manifest.permission.CAMERA),
-                        QRView.CAMERA_REQUEST_ID)
-            }
-        }
     }
 }
